+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/condition/ConditionalOnReflectRefreshTypeTest.java
@@ -60,7 +60,7 @@ public class ConditionalOnReflectRefreshTypeTest {
 				serverSocket.accept();
 			}
 			catch (IOException e) {
-				//ignore
+				e.printStackTrace();
 			}
 		}).start();
 	}